Fixed stroke loading & removed cast warning

+++ includes/menu/bar.h
@@ -20,7 +20,7 @@ inline auto mid_bottom(const Window &w, mth::Dim<int> d) -> mth::Point<int>
 
 inline auto gen_bar(Renderer &r, const TextureData &icon_map, IconDB &icons, const Icon *h = nullptr) -> Bar
 {
-	const auto bar = bar_size(icons.size());
+	const auto bar = bar_size((int)icons.size());
 	auto	   t   = r.create_texture(bar.w, bar.h);
 
 	r.set_render_target(t);
